Fix bug that prevented messages via patternchannel from getting sent
This bug occured when a user sent a message to a channel that was pattern
subscribed to. Then, the _handle_get_command function would return None as the
channel that was requested was not in the allowed channel names.
This was fixed by adding a check that returned True if the channel_name appeared
in the channel_patterns, thus passing that first test in _handle_get_command