Subscribe to specific channel matched by pattern
Enable clients to subscribe to channels by name, that were
subscribed to by pattern in the WebsocketServer.

This could be used to subscribe to all redis channels starting with
"public" in the WebsocketServer and letting web clients subscribe to
them by name (e.g. with `SUB public_healthcheck`).

Bumping the version to 0.1 since this is a new exiting feature!